requirements.txt
# Core dependencies
openai-whisper>=20231117
gradio>=4.0.0
requests>=2.31.0
soundfile>=0.12.1
numpy>=1.24.0
torch>=2.0.0
